requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
typing-extensions>=4.8.0
pyyaml>=6.0